Fixed NSNotification observer issue with LITableLayout
I'm not entirely satisfied with this fix, but it seems to eliminate an issue with deallocated layouts.

I believe the underlying problem is that by the time -[LITableLayout dealloc] is called, the the object can no longer access table components from which it needs to de-register itself as a delegate. I may revisit this whole issue when time permits.